More Collision Overhaul + started Editor Properties
Red lines in overlay are contact surfaces, soon to be used for
interactions like rotation and slopes. Little bug where they draw
diagonal though.

+++ src/entities/EntityDynamic.java
@@ -87,7 +87,7 @@ public class EntityDynamic extends EntityStatic{
     }
     
 	public Boundary getDeltaBoundary(){
-		return boundary.atPosition(x+dx,y+dy);
+		return boundary.atPosition( (int) (x+dx), (int) (y+dy ));
 	}
     
     public boolean isColliding(){ return isColliding; }
